requirements.txt
apify-client
apify
fastapi
python-dotenv
uvicorn
pandas
numpy
mysql-connector-python
cloudinary